split main.rs
