+++ src/core/mod.rs
@@ -1,5 +1,8 @@
 pub mod authenticator;
+pub mod cli;
 pub mod common;
 pub mod configuration;
+pub mod core;
 pub mod discovery;
+pub mod logger;
 pub mod prometheus;
